Add: new schema property and validateToken property for authentication

+++ controllers/contactController.cjs
@@ -3,15 +3,15 @@ const Contact = require("../models/contactModel.cjs");
 const { constants } = require("../constants.cjs");
 //@desc Get all contacts
 //@route GET/api/contacts
-//@access public
+//@access private
 const getContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find();
+  const contacts = await Contact.find({ user_id: req.user.id });
   res.status(200).json(contacts);
 });
 
 //@desc Create a contact
 //@route POST/api/contacts
-//@access public
+//@access private
 const createContact = asyncHandler(async (req, res) => {
   console.log("The request body is ", req.body);
   const { name, email, phone } = req.body;
@@ -23,6 +23,7 @@ const createContact = asyncHandler(async (req, res) => {
       name,
       email,
       phone,
+      user_id: req.user.id,
     });
     res.status(201).json(contact);
   }
@@ -30,25 +31,32 @@ const createContact = asyncHandler(async (req, res) => {
 
 //@desc Get a contact
 //@route GET/api/contacts/:id
-//@access public
+//@access private
 const getContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (contact != null) {
     res.status(200).json(contact);
   } else {
+    res.status(404);
     throw new Error("Not found!");
   }
 });
 
 //@desc Modify a contact
 //@route PUT/api/contacts/:id
-//@access public
+//@access private
 const updateContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Not found!");
   } else {
+    if (contact.user_id.toString() !== req.user.id) {
+      res.status(403);
+      throw new Error(
+        "User don't have sufficient permissions to update other user's contact."
+      );
+    }
     const updatedContact = await Contact.findByIdAndUpdate(
       req.params.id,
       req.body,
@@ -60,13 +68,19 @@ const updateContact = asyncHandler(async (req, res) => {
 
 //@desc Delete a contact
 //@route DELETE/api/contacts/:id
-//@access public
+//@access private
 const deleteContact = asyncHandler(async (req, res) => {
   const contact = await Contact.findById(req.params.id);
   if (!contact) {
     res.status(404);
     throw new Error("Not found!");
   } else {
+    if (contact.user_id.toString() !== req.user.id) {
+      res.status(403);
+      throw new Error(
+        "User don't have sufficient permissions to delete other user's contact."
+      );
+    }
     await Contact.deleteOne(contact);
     res.status(200).json(contact);
   }
